fix: by pass failed wp.org deployment
We're using `abraham/twitteroauth` `1.2.0` which requires PHP `7.x`. The PHP version shipped with Ubuntu latest (20.04.3 LTS at the time of writing) is PHP `8.0`. That's why composer throws the requirement errors. This commit fixes that issue by using `--ignore-platform-reqs` flag to bypass the requirement check. However, the best way to solve this issue is upgrading library version to more recent version.